Create change
Patch-set: 1
Change-id: I07c3444770bab33e204b95549e34f0e12d579386
Subject: This is my proposed device tree. I realize there are a lot of changes, but it is what it is. Importantly: this fixes bluetooth, and uses bcmdhd_29 instead of bcmdhd (intended for BCM4330 in our kernel). It also brings everything up-to-date with the .18 stock release. See my kernel and vendor commits as well.
Branch: refs/heads/cm-10.1
Commit: 07c3444770bab33e204b95549e34f0e12d579386