Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: This is my proposed device tree. I realize there are a lot of changes, but it is what it is. Importantly: this fixes bluetooth, and uses bcmdhd_29 instead of bcmdhd (intended for BCM4330 in our kernel). It also brings everything up-to-date with the .18 stock release. See my kernel and vendor commits as well.
Commit: 466375894eb1fc7620d4400a33deea24a0c9cdc0