Update patch set 2
Patch Set 2:

The only thing I've noticed is that the device seems to have slowed way down. Battery is off the chart. SOD is vanquished. Just slow, suddenly. Not sure if it's device or kernel. I know you mentioned the kernel heap revert. So this may be the wrong place. But this IS close to greatness.

Patch-set: 2