Update patch set 5
Patch Set 5:

John, this is because it needs the overlay/frameworks/base/core/res/res/values/config.xml to be used by DeviceHandlerService that put the real handlers in the core functions of the key and battery classes. Without that the asusdec library does nothing. When it is enabled dock key functions should work and the dock battery icon should report the battery of the dock.

Maybe this change http://review.cyanogenmod.org/#/c/31802/ helps on your search for low frequencies

Patch-set: 5