Update patch set 2
Patch Set 2:

Actually you don't configure the device as xhdpi, you rather enable xhdpi drawables while keeping mdpi as primary source. Also, no need to quote people's name who commented on the ticket as there will be a link to it on GitHub :) And no need to put patchsets difference in the commit message - put them rather in comments to keep the commit message clean for what the change actually is.

Patch-set: 2