Update patch set 1
Patch Set 1:

@Aaron,
Do you know which partition is the boot and which is the recovery partition onthe TF300T? Currently in the recovery fstab there are no entries reffering a block device partition. If i do a cat /proc/partitions i see a mmcblk0boot1 and mmcblk0boot0. I am guessing that those partitions are the boot and recovery partitions. But which is which...

Patch-set: 1