Update patch set 1
Patch Set 1:

You *do not* want to touch mmcblk0boot0/boot1. It's NOT boot/recovery - it's actually the "hidden" of the eMMC that usually holds the BCT, the tegra partition table and the bootloader. 

The asus transformer series devices does not have exposed partitions for boot and recovery, e.g. they are not in the GPT. The bootloader uses a seperate partition table to know where it is.
(That said, it's not really very relevant to this change)

Patch-set: 1